+++ lib/genet-node/src/session.cpp
@@ -209,13 +209,11 @@ NAN_METHOD(SessionWrapper::frames) {
       return;
     }
     Session *session = wrapper->event->session;
-
-    size_t length = end - start;
+    uint32_t length = end - start;
     std::vector<const Frame *> dst;
     dst.resize(length);
     genet_session_frames(session, start, end, &length, dst.data());
     dst.resize(length);
-
     auto array = Nan::New<v8::Array>(length);
     for (uint32_t index = 0; index < length; ++index) {
       array->Set(index, FrameWrapper::wrap(dst[index]));
@@ -317,7 +315,7 @@ NAN_METHOD(SessionWrapper::filteredFrames) {
     }
     Session *session = wrapper->event->session;
 
-    size_t length = end - start;
+    uint32_t length = end - start;
     std::vector<const Frame *> dst;
     dst.resize(length);
     genet_session_filtered_frames(session, id, start, end, &length, dst.data());
